dev: mediump -> highp in vertex

+++ res/shaders/font.vert
@@ -1,5 +1,6 @@
 #version 300 es
-precision mediump float;
+precision highp float;
+
 layout (location = 0) in vec4 vertex;
 out vec2 TexCoords;
